refactor: Use clearer component names
